+++ examples/src/main/java/org/partiql/examples/PartiQLCompilerPipelineAsyncJavaExample.java
@@ -16,6 +16,8 @@ import org.partiql.examples.util.Example;
 import org.partiql.lang.compiler.PartiQLCompilerAsync;
 import org.partiql.lang.compiler.PartiQLCompilerAsyncBuilder;
 import org.partiql.lang.compiler.PartiQLCompilerPipelineAsync;
+import org.partiql.lang.eval.BindingId;
+import org.partiql.lang.eval.BindingName;
 import org.partiql.lang.eval.Bindings;
 import org.partiql.lang.eval.EvaluationSession;
 import org.partiql.lang.eval.ExprValue;
@@ -63,14 +65,28 @@ public class PartiQLCompilerPipelineAsyncJavaExample extends Example {
                 .globals(globalVariables)
                 .build();
 
-        final GlobalVariableResolver globalVariableResolver = bindingName -> {
-            ExprValue value = session.getGlobals().get(bindingName);
-
-            if (value != null) {
-                return new GlobalResolutionResult.GlobalVariable(bindingName.getName());
-            }
-            else {
+        final GlobalVariableResolver globalVariableResolver = new GlobalVariableResolver() {
+            @NotNull
+            @Override
+            public GlobalResolutionResult resolveGlobal(@NotNull BindingId bindingId) {
+                // In this example, we don't allow for qualified identifiers.
+                if (!bindingId.hasQualifier()) {
+                    return resolveGlobal(bindingId.getIdentifier());
+                }
                 return GlobalResolutionResult.Undefined.INSTANCE;
+            }
+
+            @NotNull
+            @Override
+            public GlobalResolutionResult resolveGlobal(@NotNull BindingName bindingName) {
+                ExprValue value = session.getGlobals().get(bindingName);
+
+                if (value != null) {
+                    return new GlobalResolutionResult.GlobalVariable(bindingName.getName());
+                }
+                else {
+                    return GlobalResolutionResult.Undefined.INSTANCE;
+                }
             }
         };