+++ partiql-lang/src/main/kotlin/org/partiql/lang/planner/GlobalVariableResolver.kt
@@ -1,6 +1,7 @@
 package org.partiql.lang.planner
 
 import org.partiql.lang.eval.BindingCase
+import org.partiql.lang.eval.BindingId
 import org.partiql.lang.eval.BindingName
 
 /**
@@ -16,6 +17,21 @@ sealed class GlobalResolutionResult {
      * information about global variables.
      */
     data class GlobalVariable(val uniqueId: String) : GlobalResolutionResult()
+
+    /**
+     * A success case. Refers to a variable that is contained within a namespace (AKA catalog/schema).
+     */
+    abstract class NamespacedVariable(
+        val uniqueId: String
+    ) : GlobalResolutionResult() {
+        /**
+         * When attempting to resolve a qualified identifier, say `FROM cat1.schema1.table1.attr1.attr2`, the [GlobalVariableResolver]
+         * _may_ resolve only the first few steps of the identifier, say `cat1.schema1.table1`. Therefore, it must
+         * return the resolved [uniqueId] (say `cat1:::schema1:::table1`) and the remaining steps `attr1.attr2` in order for the planner
+         * to convert these to path expressions.
+         */
+        abstract fun getRemainingSteps(): List<BindingName>
+    }
 
     /** A failure case, indicates that resolution did not match any variable. */
     object Undefined : GlobalResolutionResult()
@@ -55,6 +71,18 @@ fun interface GlobalVariableResolver {
      * be used outside this project.
      */
     fun resolveGlobal(bindingName: BindingName): GlobalResolutionResult
+
+    /**
+     * Resolves a potentially qualified [BindingId] in the database environment.
+     * By default, if the [bindingId] does not have a qualified identifier passed to it, it will resolve using just
+     * the unqualified identifier. If it is qualified, it will return a [GlobalResolutionResult.Undefined].
+     */
+    fun resolveGlobal(bindingId: BindingId): GlobalResolutionResult {
+        if (!bindingId.hasQualifier()) {
+            return resolveGlobal(bindingId.getIdentifier())
+        }
+        return GlobalResolutionResult.Undefined
+    }
 
     companion object {