+++ partiql-lang/src/test/kotlin/org/partiql/lang/planner/transforms/AstToLogicalVisitorTransformTests.kt
@@ -1348,3 +1348,14 @@ class AstToLogicalVisitorTransformTests {
         )
     }
 }
+
+fun dmlTarget(node: PartiqlLogical.Identifier): PartiqlLogical.DmlTarget {
+    return PartiqlLogical.DmlTarget(
+        PartiqlLogical.TableName(
+            PartiqlLogical.IdentifierChain(
+                head = node,
+                qualifier = emptyList()
+            )
+        )
+    )
+}